refactored to have new point_sources.py file
