+++ src/components/statevector_component/aggregation.py
@@ -3,16 +3,8 @@
 
 import sys
 import yaml
-import datetime
-import requests
 import xarray as xr
 import numpy as np
-import pandas as pd
-import geopandas as gpd
-from bs4 import BeautifulSoup
-import matplotlib.pyplot as plt
-from shapely.geometry import Point
-from dateutil.relativedelta import relativedelta
 
 from src.inversion_scripts.point_sources import get_point_source_coordinates
 from src.inversion_scripts.imi_preview import (
@@ -303,115 +295,6 @@ def generate_cluster_pairs(config, sensitivities):
     return sorted(cluster_pairs, key=lambda x: x[0])
 
 
-def get_plumes(month, year):
-    """
-    Description:
-        Scrapes the SRON database for weekly methane plumes, saving each week's data as csv files
-    arguments:
-        month        String : the month (number) for which to select the plumes
-        year         String : the year (number) for which to select the plumes
-    Returns:                pd.Dataframe() : pandas dataframe with all of the plumes detected for that month
-    """
-    sron_url = "https://earth.sron.nl/wp-content/uploads/"     #URL of the SRON database for weekly methane plumes
-    url = sron_url + year + "/" + month.zfill(2)
-    response = requests.get(url)
-    parser = BeautifulSoup(response.content, 'html.parser')
-    plume = pd.DataFrame()
-    for link in parser.find_all('a'):
-        if ('.csv' in link.get('href') and 'SRON_Weekly_Methane_Plumes' in link.get('href')): #filters through all .csv files containing "SRON_Weekly_Methane_Plumes"
-            csvUrl = url + "/" + link.get('href')
-            dates = csvUrl.split("_v")[1]
-            try:
-                rcsv = requests.get(csvUrl, allow_redirects=True)
-                file = f"SRON_{dates}"
-                open(file, 'wb').write(rcsv.content) # downloads all of the plumes from that week in a CSV file in the current directory
-                df = pd.read_csv(file) # reads from the csv file into a pandas dataframe
-                plume = plume.append(df, ignore_index=True)
-            except:
-                print(f"Warning: Unable to access data for csv file at {csvUrl}. The file may not exist or there may be a connection problem.")
-    return plume
-
-
-def shapefile_filter(plumes, shapefile_path):
-    """
-    Description:
-        Removes any plumes (coordinates) that are not within a given shapefile
-    arguments:
-        plumes   pd.Dataframe() : a pandas dataframe with columns 'lon' and 'lat'
-        shapefile_path      String : a string with the path to the shapefile of the ROI
-    Returns:                pd.Dataframe() : pandas dataframe only containing coordinates within the shapefile
-    """
-    shapefile = gpd.read_file(shapefile_path)
-    for lon, lat in zip(plumes['lon'], plumes['lat']):   
-        point = Point(lon, lat)
-        is_within = shapefile.contains(point)
-        # checks if it is within any of the polygons if multiple
-        is_within_any = is_within.any()
-        if not is_within_any:
-            plumes  = plumes[(plumes['lon'] != float(lon)) | (plumes['lat'] != float(lat))]
-    
-    return plumes
-
-
-def rectangular_filter(plumes, LatMax, LatMin, LonMax, LonMin):
-    """
-    Description:
-        Removes any plumes (coordinate) not within a given set of coordinates
-    arguments:
-        plumes   pd.Dataframe() : a pandas dataframe with columns 'lon' and 'lat'
-        LatMax          float : a float indicating the maximum latitude in the ROI
-        LatMin          float : a float indicating the minimum latitude in the ROI
-        LonMax          float : a float indicating the maximum longitude in the ROI
-        LonMin          float : a float indicating the minimum longitude in the ROI
-    Returns:                pd.Dataframe() : pandas dataframe
-    """
-    inLat = (plumes['lat'] > LatMin) & (plumes['lat'] < LatMax)
-    inLon = (plumes['lon'] > LonMin) & (plumes['lon'] < LonMax)
-    filtered_plumes = plumes[inLat & inLon]
-    return filtered_plumes
-
-def SRON_plumes(config):
-    """
-    Description:
-        Selects all the recorded methane plumes on the SRON database for the selected time frame and region
-    arguments:
-        config              parsed YAML file 
-    Returns:                [[]] : list of [lat, lon] coordinates of floats
-    """
-    
-    #variables from config file, specifying time period and region.
-    plumes = pd.DataFrame()
-    shapefile_path = config["ShapeFile"] 
-    startDate = datetime.datetime.strptime(str(config["StartDate"]), "%Y%m%d")
-    endDate = datetime.datetime.strptime(str(config["EndDate"]), "%Y%m%d")
-    if endDate.year < 2023: # SRON plumes are only available beginning in 2023
-        return None
-    custom_vectorfile = not config["CreateAutomaticRectilinearStateVectorFile"]
-    LatMax = config["LatMax"]
-    LatMin = config["LatMin"]
-    LonMax = config["LonMax"]
-    LonMin = config["LonMin"]
-    currentDate = startDate
-
-    
-    
-    #calls the get_plumes function for every month in the selected time frame
-    while currentDate <= endDate:
-        if currentDate.year >= 2023: # SRON plumes are only available beginning in 2023
-            p = get_plumes(str(currentDate.month), str(currentDate.year))
-            plumes = pd.concat([plumes, pd.DataFrame(p)], ignore_index=True) 
-        currentDate = currentDate + relativedelta(months=1)
-    
-    #filters through the dataset to remove any plumes outside the ROI
-    if custom_vectorfile:
-        plumes = shapefile_filter(plumes, shapefile_path) #calls function to filter through coordinates found in shapefile
-    else:
-        plumes = rectangular_filter(plumes, LatMax, LatMin, LonMax, LonMin)
-
-    plumes_list = plumes[['lat', 'lon']].values.tolist()
-    return plumes_list
-
-
 def force_native_res_pixels(config, clusters, sensitivities):
     """
     Description:
@@ -428,7 +311,9 @@ def force_native_res_pixels(config, clusters, sensitivities):
 
     if len(coords) == 0:
         # No forced pixels inputted
-        print(f"No forced native pixels specified or in {config['PointSourceDatasets']} dataset.")
+        print(
+            f"No forced native pixels specified or in {config['PointSourceDatasets']} dataset."
+        )
         return sensitivities
 
     if config["Res"] == "0.25x0.3125":
@@ -437,24 +322,21 @@ def force_native_res_pixels(config, clusters, sensitivities):
     elif config["Res"] == "0.5x0.625":
         lat_step = 0.5
         lon_step = 0.625
-    
+
     for lat, lon in coords:
         lon = np.floor(lon / lon_step) * lon_step
         lat = np.floor(lat / lat_step) * lat_step
 
- 
-    # Remove any duplicate coordinates within the same gridcell. 
-    coords = sorted(set(map(tuple, coords)), reverse=True) 
+    # Remove any duplicate coordinates within the same gridcell.
+    coords = sorted(set(map(tuple, coords)), reverse=True)
     coords = [list(coordinate) for coordinate in coords]
-    
 
     if len(coords) > config["NumberOfElements"]:
-        coords = coords[0:config["NumberOfElements"] - 1]
+        coords = coords[0 : config["NumberOfElements"] - 1]
 
     for lat, lon in coords:
         binned_lon = np.floor(lon / lon_step) * lon_step
         binned_lat = np.floor(lat / lat_step) * lat_step
-
 
         try:
             cluster_index = int(
